fixed connection Issues & DTOs & fixed no name of RecipeIngredient
RecipeController, RecipeIngredientController, RecipeIngredient

RecipeController:
I fixed the Data transfer Issues for the Category mehtods. Here was the Issue that the wrong method was choosen to add and remove a Category

RecipeIngredientController:
Here I also fixed the Data transfer and fitted towards the frontend

RecipeIngredient:
I added a beforeInsert method to ensure that the Object has also a Name that the frontend is able to display.